#52: move from sentry/browser to sentry/react
