+++ applications/osb-portal/package.json
@@ -43,7 +43,7 @@
     "ts-loader": "^6.2.2",
     "tslint": "^6.1.2",
     "tslint-react": "^4.2.0",
-    "typescript": "^3.9.3",
+    "typescript": "^3.9.6",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.11.0",
@@ -51,7 +51,7 @@
   },
   "dependencies": {
     "@material-ui/lab": "^4.0.0-alpha.55",
-    "@sentry/browser": "^5.17.0",
+    "@sentry/react": "^5.19.1",
     "@types/markdown-it": "^10.0.1",
     "keycloak-js": "^10.0.2",
     "less-vars-to-js": "^1.3.0",
